Furnace Foam Evaluator Command-Line Version
This program is designed to compute the Furnace Foam evaluation of a dot configuration on the foam. It converts the input and output into polynomials in the cohomology ring of the representation variety of the Jumping Jack and the cohomology ring of complete flags on C^3.
Results are returned in the notation used in my paper and notes. The user can choose which notation to use. The program also has an option to record the calculations in a .txt file formatted for copy/paste into LaTeX. It can also generate all possible inputs for the furnace foam that have nonzero images. 

I am currently working on a GUI version which can take multiple inputs and return LaTeX formatted answers that can be copied and pasted. It will also generate .txt files of the results if the user wishes.