Furnace Foam Evaluator v1.0
This is the Furnace Foam Evaluator with a Graphical User Interface (GUI). It can save output to a txt file. It also has the ability to evaluate a whole list of inputs. The output is copy and paste friendly. I did this project on NetBeans, I used this as an opportunity to try it out.